parse to float for better money handling...

+++ src/functions/format-backers.js
@@ -1,7 +1,7 @@
 export default function (backers, showAll) {
     return backers
         .map(backer => (Object.assign({}, backer, {
-            contributed: parseInt(backer.contributed, 10)
+            contributed: parseFloat(backer.contributed, 10)
         })))
         .filter( backer => ( showAll ? true : backer.contributed > 0))
         .sort((a, b) => (b.contributed - a.contributed));
